Not legal TypeScript syntax
The parameter list (for a lambda) must include a variable name.

Quoting from typescriptlang.org: 
https://www.typescriptlang.org/docs/handbook/functions.html
"A function’s type has the same two parts: the type of the arguments and the return type. When writing out the whole function type, both parts are required. We write out the parameter types just like a parameter list, giving each parameter a name and a type. This name 
is just to help with readability."